Change positiveintiger field to bigintigerfield, and fix model field issue

+++ backend/rest/models.py
@@ -15,7 +15,7 @@ class Product(models.Model):
 
 
 class ProductDetail(models.Model):
-    id = models.PositiveIntegerField(primary_key=True, auto_created=False)
+    id = models.BigIntegerField(primary_key=True, auto_created=False)
     image = models.TextField(null=True)
     name = models.CharField(max_length=200)
